Fix always-true comparison warning

+++ src/listing.c
@@ -970,7 +970,7 @@ compute_maxes(void)
 		}
 
 		if (prop_fields.links == 1) {
-			t = DIGINUM(file_info[i].linkn);
+			t = DIGINUM((unsigned int)file_info[i].linkn);
 			if (t > maxes.links)
 				maxes.links = t;
 		}
